modify LdbcQuery10 to take only 1 month parameter (start month) rather than 2 (start and end month), because end month is simply: (start_month + 1)  % 12 change similarity score parameter in LdbcQuery10Result from double to int, because the query has changes such that the score can never not be a whole number update all tests according to these changes

+++ src/test/java/com/ldbc/driver/workloads/ldbc/snb/interactive/LdbcSnbOperationResultEqualityTest.java
@@ -312,14 +312,14 @@ public class LdbcSnbOperationResultEqualityTest {
         long personId1 = 1;
         String personFirstName1 = "2";
         String personLastName1 = "3";
-        double commonInterestScore1 = 4d;
+        int commonInterestScore1 = 4;
         String personGender1 = "5";
         String personCityName1 = "6";
 
         long personId2 = 7;
         String personFirstName2 = "8";
         String personLastName2 = "9";
-        double commonInterestScore2 = 10d;
+        int commonInterestScore2 = 10;
         String personGender2 = "11";
         String personCityName2 = "12";
